refactor: rename icrc.idl.constants
